requirements.txt
# Discord
discord.py==2.3.2
python-dotenv==1.0.0

# AWS
boto3==1.34.34

# GitHub
PyGithub==2.1.1

# Data handling
python-dateutil==2.8.2

# HTTP
aiohttp==3.9.1

# Async utilities
asyncio==3.4.3

# Webhook server
Flask==3.0.0
